[Refactor]: 스페이스 즐겨찾기 추가 API URL 엔드포인트 변경.
