--- src/main/java/com/tenten/linkhub/domain/space/controller/dto/favorite/SpaceRegisterInFavoriteApiRequest.java
@@ -1,4 +1,0 @@
-package com.tenten.linkhub.domain.space.controller.dto.favorite;
-
-public record SpaceRegisterInFavoriteApiRequest(Long spaceId) {
-}
