+++ src/main/java/com/tenten/linkhub/domain/space/controller/dto/space/SpaceRegisterInFavoriteApiResponse.java
@@ -1,4 +1,4 @@
-package com.tenten.linkhub.domain.space.controller.dto.favorite;
+package com.tenten.linkhub.domain.space.controller.dto.space;
 
 import com.tenten.linkhub.domain.space.service.dto.space.SpaceRegisterInFavoriteResponse;
